standarize exporting in all modules
