+++ packages/sql/src/index.ts
@@ -1,3 +1,3 @@
 import { SQL } from './Client';
 
-export default { SQL };
+export { SQL };
